NFC: Move ContractOpInfo to a new file in Codegen/Utils/
This prepares it to be shared by other parts in the codebase.
