+++ compiler/src/iree/compiler/Codegen/Common/GPU/CMakeLists.txt
@@ -113,6 +113,7 @@ iree_cc_library(
     iree::compiler::Codegen::Dialect::GPU::IR::IREEGPUDialect
     iree::compiler::Codegen::Transforms
     iree::compiler::Codegen::Utils
+    iree::compiler::Codegen::Utils::VectorOpUtils
     iree::compiler::Dialect::HAL::IR
   PUBLIC
 )
